% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_obis_date.R
\name{format_obis_date}
\alias{format_obis_date}
\title{Extract separate year, month and day information from obis data}
\usage{
format_obis_date(obis_data)
}
\arguments{
\item{obis_data}{is a data.frame of OBIS occurrence data 
obtained using robis::occurrence or merpWS::plot_obis}
}
\description{
The function takes an OBIS occurrence data set and
uses its eventDate columns to extract year, month
and day information.
}
\details{
\code{format_obis_date} formats the "eventDate" column
in the data frame returned by robis::occurrence function.
The eventDate entry is of the form yyyy-mm-dd hh:mm:ss,
a format that needs splitting into year, month and day
information to be useful.
}

